requirements.txt
python-telegram-bot[async]==21.7
httpx==0.27.0
python-dotenv==1.0.1
